+++ ApexShop.Infrastructure/Entities/Product.cs
@@ -7,7 +7,7 @@ public class Product
     public string? Description { get; set; }
     public decimal Price { get; set; }
     public short Stock { get; set; } // smallint: -32,768 to 32,767 (most products won't exceed this)
-    public short CategoryId { get; set; } // smallint for PostgreSQL compatibility
+    public int CategoryId { get; set; }
     public DateTime CreatedDate { get; set; }
     public DateTime? UpdatedDate { get; set; }
     public bool IsActive { get; set; } = true;
